+++ lib/compony/manage_intents_dsl.rb
@@ -12,11 +12,13 @@ module Compony
     # Adds or replaces an intent to those exposed by the component based on the intent name (override the name if you need to avoid a naming collision).
     # Intents specified this way can be retrieved and rendered by the parent component or by calling `root_intents` in case of standalone access.
     # @param [Symbol] before If specified, will insert the intent before the other.  When replacing, an element keeps its position unless `before:`` is passed.
-    def add(*, before: nil, **)
-      intent = Compony.intent(*, **@intent_opts, **)
+    def add(*args, before: nil, **kwargs)
+      intent = Compony.intent(*args, **@intent_opts, **kwargs)
       @exposed_intents.natural_push(intent.name, intent, before:)
-    rescue NameError # Ignore if the component is not actually defined
-      return nil
+    rescue NameError => e # Ignore if the component is not actually defined
+      Rails.logger.debug do
+        "Skipping intent for arguments #{args.inspect}, #{kwargs.inspect}: #{e.inspect}."
+      end
     end
 
     # DSL method