+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# unreleased
+
+- Produce debug output when skipping an intent due to a NameError
+
 # 0.10.1
 
 - Implement intent to always return the path it was initialized with in case the argument was given